Explicit unboxing
git-svn-id: https://svn.apache.org/repos/asf/commons/proper/collections/trunk@1436075 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/commons/collections/trie/ByteKeyAnalyzer.java
@@ -66,7 +66,7 @@ public class ByteKeyAnalyzer extends AbstractKeyAnalyzer<Byte> {
      * {@inheritDoc}
      */
     public boolean isBitSet(final Byte key, final int bitIndex, final int lengthInBits) {
-        return (key & mask(bitIndex)) != 0;
+        return (key.intValue() & mask(bitIndex)) != 0;
     }
 
     /**
